pyproject.toml
[tool.poetry]
name = "repl_python3_Burgir"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
ffmpeg = "^1.4"
discord-interactive = "^4.0"
"discord.py" = "^1.7.3"
discord-musicbot = "^0.0.3-alpha.20170604"
youtube_dl = "^2021.6.6"
PyNaCl = "^1.4.0"
replit = "^3.2.4"
lockfile = "^0.12.2"
label = "^0.1.1"
urllib3_1_26_2 = "^1.26.2"
simplejson = "^3.17.5"
idna = "^3.2"
urlparse2 = "^1.1.1"
asyncio-cancel-token = "^0.1.0a2"
pip = "^21.2.4"
pyshorteners = "^1.0.1"
Flask = "^2.0.1"
meme_get = "^1.0.2-beta.1"
praw = "^7.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
